# internal/api/openapi.yaml
openapi: 3.0.3
info:
  title: k8s-ui API
  description: GitOps dashboard control-plane API.
  version: 0.1.0
servers:
  - url: /
security:
  - bearerAuth: []
paths:
...
  /api/v1/applications/{name}:
    parameters:
      - in: path
        name: name
        required: true
        schema: { type: string }
    get:
      summary: Get an application by name.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Application" }
        "404":
          $ref: "#/components/responses/Error"
    put:
      summary: Update an application.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateApplicationRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Application" }
    delete:
      summary: Delete an application.
      responses:
        "204": { description: No Content }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema:
            type: object
            properties:
              error:   { type: string }
              message: { type: string }
  schemas:
    Application:
      type: object
      required: [name, source, destination, status, createdAt, updatedAt]
      properties:
        name: { type: string }
        project: { type: string }
        source: { $ref: "#/components/schemas/AppSource" }
        destination: { $ref: "#/components/schemas/AppDestination" }
        syncPolicy: { $ref: "#/components/schemas/SyncPolicy" }
        status: { $ref: "#/components/schemas/AppStatus" }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    AppSource:
      type: object
      properties:
        repoUrl: { type: string }
        path: { type: string }
        targetRevision: { type: string }
        helmValues:
          type: object
          additionalProperties: true
          description: Optional values merged via helm template -f when path is a Helm chart
    AppDestination:
      type: object
      properties:
        cluster: { type: string }
        namespace: { type: string }
    SyncPolicy:
      type: object
      properties:
        automated:
          type: object
          nullable: true
          properties:
            prune: { type: boolean }
            selfHeal: { type: boolean }
        syncOptions:
          type: array
          items: { type: string }
          description: Argo-style options (e.g. CreateNamespace=true); subset interpreted by k8s-ui
        managedNamespaceMetadata:
          type: object
          properties:
            labels:
              type: object
              additionalProperties: { type: string }
            annotations:
              type: object
              additionalProperties: { type: string }
        createNamespace: { type: boolean }
        materializeChildApps: { type: boolean }
    AppStatus:
      type: object
      properties:
        sync: { type: string }
        health: { type: string }
        observedRevision: { type: string }
        lastSyncedAt: { type: string, format: date-time, nullable: true }
        message: { type: string }
        syncOperation:
          type: object
          nullable: true
          properties:
            status: { type: string, description: "Pending or Running" }
            message: { type: string }
        lastCompletedSync:
          type: object
          nullable: true
          properties:
            status: { type: string, description: "Succeeded or Failed" }
            message: { type: string }
...
    UpdateApplicationRequest:
      type: object
      properties:
        source: { $ref: "#/components/schemas/AppSource" }
        destination: { $ref: "#/components/schemas/AppDestination" }
        syncPolicy: { $ref: "#/components/schemas/SyncPolicy" }
...